revisi

+++ app/controllers/Admins.php
@@ -4,6 +4,13 @@ class Admins extends Controller
     public $response = [];
     public function index()
     {
+        if ($_SERVER['REQUEST_METHOD'] !== 'GET') {
+            http_response_code(405);
+            header('Content-Type: application/json');
+            echo json_encode(['error' => 'Method Not Allowed']);
+            return;
+        }
+        
         $data['judul'] = 'Menu Admin - Admins';
         $data['admins'] = $this->model('admins_model')->loadAdmin();
         header('Content-Type: application/json');
@@ -15,32 +22,47 @@ class Admins extends Controller
 
     public function addAdmin()
     {
-        $emails = $this->model('admins_model')->checkEmailAdmin($_POST['email']);
+        if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+            http_response_code(405);
+            header('Content-Type: application/json');
+            echo json_encode(['error' => 'Method Not Allowed']);
+            return;
+        }
+
+        $json_data = file_get_contents("php://input");
+        $data = json_decode($json_data, true);
+
+        if ($data['pass'] !== $data['confirmPass']) {
+            $response['status'] = 'error';
+            $response['message'] = 'Password not match';
+            header('Content-type: application/json');
+            echo json_encode($response);
+            exit;
+        }
+
+        $emails = $this->model('admins_model')->checkEmailAdmin($data['email']);
         if ($emails) {
             $response['status'] = 'error';
             $response['message'] = 'Email is already registered';
-            $response['http_code'] = 400;
         } else {
-            if ($this->model('admins_model')->addAdmin($_POST) > 0) {
+            if ($this->model('admins_model')->addAdmin($data) > 0) {
                 $response['status'] = 'success';
                 $response['message'] = 'Admin added successfully';
                 $response['http_code'] = 201;
             } else {
                 $response['status'] = 'error';
                 $response['message'] = 'Failed to add admin';
-                $response['http_code'] = 500;
             }
         }
 
-        http_response_code($response['http_code']);
-        unset($response['http_code']);
         header('Content-Type: application/json');
         echo json_encode($response);
     }
-    public function edit()
+
+    public function edit(string $id)
     {
         if ($_SERVER['REQUEST_METHOD'] !== 'PUT' && $_SERVER['REQUEST_METHOD'] !== 'PATCH') {
-            http_response_code(405);
+            // http_response_code(405);
             header('Content-Type: application/json');
             echo json_encode(['error' => 'Method Not Allowed']);
             return;
@@ -52,6 +74,8 @@ class Admins extends Controller
             unset($_PUT['pass']);
         }
 
+        $_PUT['id_user'] = $id;
+
         if ($this->model('admins_model')->edit($_PUT) > 0) {
             http_response_code(201);
             header('Content-Type: application/json');
@@ -60,7 +84,7 @@ class Admins extends Controller
                 'message' => 'Admin edited successfully'
             ]);
         } else {
-            http_response_code(500);
+            // http_response_code(500);
             header('Content-Type: application/json');
             echo json_encode([
                 'status' => 'error',
@@ -71,6 +95,14 @@ class Admins extends Controller
 
     public function delete($id_user)
     {
+
+        if ($_SERVER['REQUEST_METHOD'] !== 'DELETE') {
+            http_response_code(405);
+            header('Content-Type: application/json');
+            echo json_encode(['error' => 'Method Not Allowed']);
+            return;
+        }
+        
         if ($this->model('admins_model')->delete($id_user) > 0) {
             http_response_code(201);
             header('Content-Type: application/json');
