+++ app/controllers/Customers.php
@@ -3,6 +3,13 @@ class Customers extends Controller
 {
     public function index()
     {
+        if ($_SERVER['REQUEST_METHOD'] !== 'GET') {
+            http_response_code(405);
+            header('Content-Type: application/json');
+            echo json_encode(['error' => 'Method Not Allowed']);
+            return;
+        }
+
         $data['judul'] = 'Menu Admin - Customers';
         $data['users'] = $this->model('customers_model')->loadUser();
         header('Content-Type: application/json');
@@ -14,33 +21,47 @@ class Customers extends Controller
 
     public function addUser()
     {
-        $emails = $this->model('customers_model')->checkEmailUser($_POST['email']);
+        if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+            http_response_code(405);
+            header('Content-Type: application/json');
+            echo json_encode(['error' => 'Method Not Allowed']);
+            return;
+        }
+
+        $json_data = file_get_contents("php://input");
+        $data = json_decode($json_data, true);
+
+        if ($data['pass'] !== $data['confirmPass']) {
+            $response['status'] = 'error';
+            $response['message'] = 'Password not match';
+            header('Content-type: application/json');
+            echo json_encode($response);
+            exit;
+        }
+
+        $emails = $this->model('customers_model')->checkEmailUser($data['email']);
         if ($emails) {
             $response['status'] = 'error';
             $response['message'] = 'Email is already registered';
-            $response['http_code'] = 400;
         }
 
-        if ($this->model('customers_model')->addUser($_POST) > 0) {
+        if ($this->model('customers_model')->addUser($data) > 0) {
             $response['status'] = 'success';
             $response['message'] = 'Admin added user successfully';
-            $response['http_code'] = 201;
         } else {
             $response['status'] = 'error';
             $response['message'] = 'Failed to added user';
             $response['http_code'] = 500;
         }
-
-        http_response_code($response['http_code']);
-        unset($response['http_code']);
+        
         header('Content-Type: application/json');
         echo json_encode($response);
     }
-    public function edit()
+    public function edit(string $id)
     {
 
-        if ($_SERVER['REQUEST_METHOD'] !== 'PUT' && $_SERVER['REQUEST_METHOD'] !== 'PATCH') {
-            http_response_code(405);
+        if ($_SERVER['REQUEST_METHOD'] !== 'PUT') {
+            // http_response_code(405);
             header('Content-Type: application/json');
             echo json_encode(['error' => 'Method Not Allowed']);
             return;
@@ -52,7 +73,9 @@ class Customers extends Controller
             unset($_PUT['pass']);
         }
 
-        if ($this->model('customers_model')->edit($_POST) > 0) {
+        $_PUT['id_user'] = $id;
+
+        if ($this->model('customers_model')->edit($_PUT) > 0) {
             http_response_code(201);
             header('Content-Type: application/json');
             echo json_encode([
@@ -60,7 +83,7 @@ class Customers extends Controller
                 'message' => 'Admin edited user successfully'
             ]);
         } else {
-            http_response_code(500);
+            // http_response_code(500);
             header('Content-Type: application/json');
             echo json_encode([
                 'status' => 'error',
@@ -71,6 +94,13 @@ class Customers extends Controller
 
     public function delete($id_user)
     {
+        if ($_SERVER['REQUEST_METHOD'] !== 'DELETE') {
+            http_response_code(405);
+            header('Content-Type: application/json');
+            echo json_encode(['error' => 'Method Not Allowed']);
+            return;
+        }
+
         if ($this->model('customers_model')->delete($id_user) > 0) {
             http_response_code(201);
             header('Content-Type: application/json');
